// apps/web-svelte/src/lib/api/generated/model/bookAudioChapter.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface BookAudioChapter {
  audioTrackId: string;
  /** @nullable */
  audioMarkerId: string | null;
  title: string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$ */
  startSeconds: number | string;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
     */
  endSeconds: number | string | null;
}